Implemented  issue Seed the database with users and categories with a run error
