+++ budget-manager-server/src/BudgetManager.Infrastructure/Data/Extensions/DatabaseExtensions.cs
@@ -7,14 +7,13 @@ public static class DatabaseExtensions
     public static async Task SeedAsync(ApplicationDbContext context)
     {
         await SeedUsersAsync(context);
-        await context.SaveChangesAsync();
-
         await SeedCategoriesAsync(context);
+        await context.SaveChangesAsync();
     }
 
     private static async Task SeedUsersAsync(ApplicationDbContext context)
     {
-        if (!context.Users.Any())
+        if (!await context.Users.AnyAsync())
         {
             await context.Users.AddRangeAsync(InitialData.Users);
             await context.SaveChangesAsync();
@@ -23,19 +22,10 @@ public static class DatabaseExtensions
 
     private static async Task SeedCategoriesAsync(ApplicationDbContext context)
     {
-        if (!context.Categories.Any())
+        if (!await context.Categories.AnyAsync())
         {
             await context.Categories.AddRangeAsync(InitialData.Categories);
             await context.SaveChangesAsync();
         }
-    }
-
-    public static async Task InitializeDatabaseAsync(this WebApplication app)
-    {
-        using var scope = app.Services.CreateScope();
-        var context = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
-
-        await context.Database.MigrateAsync();
-        await DatabaseExtensions.SeedAsync(context);
     }
 }
