+++ budget-manager-server/src/BudgetManager.Infrastructure/DependencyInjection.cs
@@ -9,6 +9,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.OpenApi.Models;
 using System.Text;
 using System.Reflection;
+using BudgetManager.Infrastructure.Data.Extensions;
 
 namespace BudgetManager.Infrastructure;
 
@@ -136,7 +137,6 @@ public static class DependencyInjection
         return services;
     }
 
-
     private static IServiceCollection AddDatabase(this IServiceCollection services, IConfiguration configuration)
     {
         services.AddDbContext<ApplicationDbContext>(options =>
@@ -145,5 +145,14 @@ public static class DependencyInjection
         services.AddScoped<IApplicationDbContext, ApplicationDbContext>();
 
         return services;
+    }
+
+    public static async Task InitializeDatabaseAsync(this WebApplication app)
+    {
+        using var scope = app.Services.CreateScope();
+        var context = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
+
+        await context.Database.MigrateAsync();
+        await DatabaseExtensions.SeedAsync(context);
     }
 }
